+++ server/build.gradle
@@ -20,7 +20,7 @@ repositories {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
+    //implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     compileOnly 'org.projectlombok:lombok'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
@@ -28,12 +28,18 @@ dependencies {
     runtimeOnly 'mysql:mysql-connector-java'
     annotationProcessor 'org.projectlombok:lombok'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.security:spring-security-test'
+    //testImplementation 'org.springframework.security:spring-security-test'
 
-    implementation group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
-    implementation group: 'io.openvidu', name: 'openvidu-java-client', version: '2.20.0'
+    // Spring Validation
+    implementation 'org.springframework.boot:spring-boot-starter-validation:2.6.2'
+
+    // https://mvnrepository.com/artifact/io.openvidu/openvidu-java-client
+    implementation 'io.openvidu:openvidu-java-client:2.2.0'
+
 }
 
 tasks.named('test') {
     useJUnitPlatform()
 }
+
+
